add identity server

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -6,9 +6,9 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Infrastructure;
 
 namespace AuctionService.Controllers;
 
@@ -53,12 +53,14 @@ public class AuctionsController : ControllerBase
     return _mapper.Map<AuctionDto>(auction);
   }
 
+  [Authorize]
   [HttpPost]
   public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
   {
     var auction = _mapper.Map<Auction>(auctionDto);
-    //Todo add current user as seller
-    auction.Seller = "test";
+
+    auction.Seller = User.Identity.Name;
+
     _context.Auctions.Add(auction);
 
     var newAuction = _mapper.Map<AuctionDto>(auction);
@@ -72,6 +74,7 @@ public class AuctionsController : ControllerBase
     return CreatedAtAction(nameof(GetAuctionById), new { auction.Id }, _mapper.Map<AuctionDto>(auction));
   }
 
+  [Authorize]
   [HttpPut("{id}")]
   public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDtos updateAuctionDtos)
   {
@@ -80,7 +83,7 @@ public class AuctionsController : ControllerBase
 
     if (auction == null) return NotFound();
 
-    // TODO: check seller == username
+    if (auction.Seller != User.Identity.Name) return Forbid();
 
     auction.Item.Make = updateAuctionDtos.Make ?? auction.Item.Make;
     auction.Item.Model = updateAuctionDtos.Model ?? auction.Item.Model;
@@ -96,6 +99,8 @@ public class AuctionsController : ControllerBase
     return BadRequest("Problem saving changes");
   }
 
+
+  [Authorize]
   [HttpDelete("{id}")]
   public async Task<ActionResult> DeleteAuction(Guid id)
   {
@@ -103,9 +108,9 @@ public class AuctionsController : ControllerBase
 
     if (auction == null) return NotFound();
 
-    // todo: check seller == username
+    if (auction.Seller != User.Identity.Name) return Forbid();
 
-    await _publishEndpoint.Publish(new { Id = auction.Id.ToString() });
+    await _publishEndpoint.Publish<AuctionDeleted>(new { Id = auction.Id.ToString() });
 
     _context.Auctions.Remove(auction);
